added devise, implemented coupons

+++ app/models/price.rb
@@ -2,7 +2,7 @@ class Price
   include Mongoid::Document
   include Mongoid::Timestamps
 
-  field :amount, :type => Float, :default => 0
+  field :amount, :type => Integer, :default => 0
 
   belongs_to :product
   belongs_to :scent
